requirements.txt
praw
python-dotenv
pandas
numpy
scikit-learn
nltk
